data-pipeline/requirements.txt
pdfplumber
langchain
langchain-text-splitters
openai
pinecone
python-dotenv
